Move test environment to separate package
