+++ java-tests/validating/src/test/kotlin/io/spine/test/options/goes/GoesOneWayITest.kt
@@ -27,6 +27,9 @@
 package io.spine.test.options.goes
 
 import com.google.protobuf.Message
+import io.spine.test.options.goes.given.newBuilder
+import io.spine.test.options.goes.given.protoDescriptor
+import io.spine.test.options.goes.given.protoValue
 import io.spine.validate.ValidationException
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.assertDoesNotThrow
@@ -47,7 +50,7 @@ import org.junit.jupiter.params.provider.MethodSource
 @DisplayName("`(goes)` constraint should")
 internal class GoesOneWayITest {
 
-    @MethodSource("io.spine.test.options.goes.TestDataOneWay#onlyTargetFields")
+    @MethodSource("io.spine.test.options.goes.given.GoesOneWayTestEnv#onlyTargetFields")
     @ParameterizedTest(name = "throw if only the target `{1}` field is set")
     fun throwIfOnlyTargetFieldSet(message: Class<Message>, fieldName: String, fieldValue: Any) {
         val descriptor = message.protoDescriptor()
@@ -60,7 +63,7 @@ internal class GoesOneWayITest {
         }
     }
 
-    @MethodSource("io.spine.test.options.goes.TestDataOneWay#onlyCompanionFields")
+    @MethodSource("io.spine.test.options.goes.given.GoesOneWayTestEnv#onlyCompanionFields")
     @ParameterizedTest(name = "pass if only the companion `{1}` field is set")
     fun passIfOnlyCompanionFieldSet(
         message: Class<out Message>,
@@ -78,7 +81,7 @@ internal class GoesOneWayITest {
     }
 
     @Suppress("MaxLineLength") // So not to wrap the test name.
-    @MethodSource("io.spine.test.options.goes.TestDataOneWay#bothTargetAndCompanionFields")
+    @MethodSource("io.spine.test.options.goes.given.GoesOneWayTestEnv#bothTargetAndCompanionFields")
     @ParameterizedTest(name = "pass if both the target `{1}` and its companion `{3}` fields are set")
     fun passIfBothTargetAndCompanionFieldsSet(
         message: Class<out Message>,
